refactor(web): extract post-review prompt coordination
